Fixes several issues, including issue #13 on Github.
Note that the Github issue is one I encountered as well. The problem
appears to be caused by which-key, and thus, may not be a treeclimber
issue at all, but because which-key is so ubiquitous, it probably makes
sense to put a small workaround in treeclimber, pending a more complete
fix.

In addition to that fatal error, there were a number of maps that just
didn't work correctly, at least not in Lua files. A not necessarily
exhaustive list is presented below...

select_grow_forward/backward were broken by the presence of co-located
nodes (i.e., distinct nodes with the same extents, one a wrapper for the
other). The problem was that largest_named_descendant_for_range()
selected the outermost co-located node, with the result that the
children of the innermost weren't considered as candidates by
get_covering_nodes().

select_shrink was broken because of issues with the history stack. For
one thing, nodes were pushed onto the stack by select_expand, but *not*
by select_current_node. Thus, for a sequence like....
        select_current_node
        select_expand
        select_shrink
...the originally selected node was not in the stack, and thus was not
considered as a possible target for the shrink. Even worse, target nodes
were not required to be below the current node, which meant that the
shrink selected the *current* node in this case, making it effectively a
no-op.

Refactored the node history stack to fix select_shrink.

Originally, nodes were pushed to the stack whenever nodes were visited,
without regard to the stack's current contents or the type of movement
(ascent, descent or lateral). This led to unexpected behavior from
select_shrink, which, whenever possible, relied on the stack to find the
shrink target.

Another problem with the original approach is that the docs state that a
shrink should undo a previous expand, but this was not the case when
multiple nodes were selected prior to the expand. Interestingly,
*ranges*, not nodes, were being pushed to the stack, but the logic
forced selection of a single node as the shrink target.

In the new approach, we push to the stack only on ascent, and when
pulling from the stack, translate a multi-node range to a multi-node
shrink target. The stack is cleared whenever we descend through a child
that's outside the range at the top of the stack.
Note: Selecting the top level node pushes all the ascended-through
nodes, including the initial range/node.