Enhanced select_shrink logic.
If there are nodes on the stack and user moves cursor manually to an ancestor of
the highest stack node, this change ensures that descent with shrink will
traverse the nodes that would have been pushed to stack had treeclimber
(rather than non-treeclimber movements) been used for the ascent.

Additionally, when starting point of a shrink is a multi-select
sequence, a descendant of *any* of the selected nodes is a potentially
valid shrink target, with fallback to the first node in the set.

Finally, removed the automatic stack clear from the expand selection
functions, since expanding the selection shouldn't invalidate descendant
ranges already on the stack.

Added several useful Range helper functions.