Implement more robust fix for "Failed to resume visual mode" issue.
For background, see Issue #13 on upstream (Dkendal) repository.
Note: An earlier commit introduced an incomplete workaround, which eliminated
the error in the most common case, but not in the scenario in which previous (or
current) visual mode was linewise.

The original approach, which adjusted marks '< and '>, then used normal gv to
make the selection, was broken by the 'which-key' plugin, whose ModeChanged
logic seemed to "eat" the gv. Complicating matters were idiosyncrasies
pertaining to the visualmode() builtin, which didn't seem to be accounted for in
the original resume_visual_charwise() function. For one thing, visualmode()
doesn't return the *current* mode when visual mode is still active (unless you
implicitly end visual mode by providing an optional true argument). In any case,
while it might be possible to devise a correct approach without abandoning the
use of gv, there is a much simpler and more straightforward approach, which has
been implemented in visually_select_range().